Renamed package and main class to Updatotron
