+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.steamclock.versioncheckkotlin"
+        applicationId "com.steamclock.updatotron"
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 1
@@ -33,7 +33,7 @@ android {
 }
 
 dependencies {
-    implementation project(':versioncheckkotlin')
+    implementation project(':updatotron')
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation "com.google.android.material:material:$material"
