+++ app/src/main/AndroidManifest.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.steamclock.versioncheckkotlinsample">
+    package="com.steamclock.updatotronsample">
 
     <application
-        android:name=".App"
+        android:name="com.steamclock.updatotronsample.App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.VersionCheckKotlin">
-        <activity android:name="com.steamclock.versioncheckkotlinsample.MainActivity">
+        android:theme="@style/Theme.Updatotron">
+        <activity android:name="com.steamclock.updatotronsample.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
