+++ app/src/main/java/com/steamclock/updatotronsample/MainActivity.kt
@@ -1,0 +1,11 @@
+package com.steamclock.updatotronsample
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
