--- app/src/main/java/com/steamclock/versioncheckkotlinsample/MainActivity.kt
@@ -1,11 +1,0 @@
-package com.steamclock.versioncheckkotlinsample
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
