+++ settings.gradle
@@ -1,3 +1,3 @@
-include ':versioncheckkotlin'
+include ':updatotron'
 include ':app'
 rootProject.name = "Version Check Kotlin"
