+++ updatotron/src/main/java/com/steamclock/updatotron/UpdatotronConfig.kt
@@ -1,0 +1,10 @@
+package com.steamclock.updatotron
+
+import com.steamclock.updatotron.interfaces.*
+
+data class UpdatotronConfig(
+    val packageDetails: PackageDetails,
+    val url: String,
+    val urlFetcher: URLFetcher = NetworkURLFetcher,
+    val versionDataConverter: VersionDataConverter = DefaultVersionDataConverter,
+)
