+++ updatotron/src/main/java/com/steamclock/updatotron/models/DisplayState.kt
@@ -1,0 +1,9 @@
+package com.steamclock.updatotron.models
+
+sealed class DisplayState {
+    object Clear: DisplayState()
+    object SuggestUpdate: DisplayState()
+    object ForceUpdate: DisplayState()
+    object DownForMaintenance: DisplayState()
+    class DevelopmentFailure(val reason: String): DisplayState()
+}
