+++ updatotron/src/main/java/com/steamclock/updatotron/models/Status.kt
@@ -1,0 +1,8 @@
+package com.steamclock.updatotron.models
+
+sealed class Status {
+    object Unknown: Status()
+    object FetchFailure: Status()
+    object VersionAllowed: Status()
+    object VersionDisallowed: Status()
+}
