--- versioncheckkotlin/src/main/java/com/steamclock/versioncheckkotlin/VersionCheckConfig.kt
@@ -1,10 +1,0 @@
-package com.steamclock.versioncheckkotlin
-
-import com.steamclock.versioncheckkotlin.interfaces.*
-
-data class VersionCheckConfig(
-    val packageDetails: PackageDetails,
-    val url: String,
-    val urlFetcher: URLFetcher = NetworkURLFetcher,
-    val versionDataConverter: VersionDataConverter = DefaultVersionDataConverter,
-)
