--- versioncheckkotlin/src/main/java/com/steamclock/versioncheckkotlin/models/DisplayState.kt
@@ -1,9 +1,0 @@
-package com.steamclock.versioncheckkotlin.models
-
-sealed class DisplayState {
-    object Clear: DisplayState()
-    object SuggestUpdate: DisplayState()
-    object ForceUpdate: DisplayState()
-    object DownForMaintenance: DisplayState()
-    class DevelopmentFailure(val reason: String): DisplayState()
-}
