--- versioncheckkotlin/src/main/java/com/steamclock/versioncheckkotlin/models/Status.kt
@@ -1,8 +1,0 @@
-package com.steamclock.versioncheckkotlin.models
-
-sealed class Status {
-    object Unknown: Status()
-    object FetchFailure: Status()
-    object VersionAllowed: Status()
-    object VersionDisallowed: Status()
-}
